+++ _includes/samples/Keepers/KeepersCounter.sol
@@ -13,31 +13,43 @@ import "@chainlink/contracts/src/v0.8/KeeperCompatible.sol";
 
 contract Counter is KeeperCompatibleInterface {
     /**
-    * Public counter variable
-    */
-    uint public counter;
+     * Public counter variable
+     */
+    uint256 public counter;
 
     /**
-    * Use an interval in seconds and a timestamp to slow execution of Upkeep
-    */
-    uint public immutable interval;
-    uint public lastTimeStamp;
+     * Use an interval in seconds and a timestamp to slow execution of Upkeep
+     */
+    uint256 public immutable interval;
+    uint256 public lastTimeStamp;
 
-    constructor(uint updateInterval) {
-      interval = updateInterval;
-      lastTimeStamp = block.timestamp;
+    constructor(uint256 updateInterval) {
+        interval = updateInterval;
+        lastTimeStamp = block.timestamp;
 
-      counter = 0;
+        counter = 0;
     }
 
-    function checkUpkeep(bytes calldata /* checkData */) external view override returns (bool upkeepNeeded, bytes memory /* performData */) {
+    function checkUpkeep(
+        bytes calldata /* checkData */
+    )
+        external
+        view
+        override
+        returns (
+            bool upkeepNeeded,
+            bytes memory /* performData */
+        )
+    {
         upkeepNeeded = (block.timestamp - lastTimeStamp) > interval;
         // We don't use the checkData in this example. The checkData is defined when the Upkeep was registered.
     }
 
-    function performUpkeep(bytes calldata /* performData */) external override {
+    function performUpkeep(
+        bytes calldata /* performData */
+    ) external override {
         //We highly recommend revalidating the upkeep in the performUpkeep function
-        if ((block.timestamp - lastTimeStamp) > interval ) {
+        if ((block.timestamp - lastTimeStamp) > interval) {
             lastTimeStamp = block.timestamp;
             counter = counter + 1;
         }
