+++ _includes/samples/PriceFeeds/HistoricalPriceConsumer.sol
@@ -10,7 +10,6 @@ import "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
  */
 
 contract HistoricalPriceConsumerV3 {
-
     AggregatorV3Interface internal priceFeed;
 
     /**
@@ -19,7 +18,9 @@ contract HistoricalPriceConsumerV3 {
      * Address:	0xD4a33860578De61DBAbDc8BFdb98FD742fA7028e
      */
     constructor() {
-        priceFeed = AggregatorV3Interface(0xD4a33860578De61DBAbDc8BFdb98FD742fA7028e);
+        priceFeed = AggregatorV3Interface(
+            0xD4a33860578De61DBAbDc8BFdb98FD742fA7028e
+        );
     }
 
     /**
@@ -36,9 +37,9 @@ contract HistoricalPriceConsumerV3 {
     function getHistoricalPrice(uint80 roundId) public view returns (int256) {
         (
             uint80 id,
-            int price,
-            uint startedAt,
-            uint timeStamp,
+            int256 price,
+            uint256 startedAt,
+            uint256 timeStamp,
             uint80 answeredInRound
         ) = priceFeed.getRoundData(roundId);
         require(timeStamp > 0, "Round not complete");
